filtering clouds

+++ global_register.h
@@ -24,12 +24,12 @@ class GlobalDQReg {
 
   private:
     std::vector<pcl::PointCloud<Point>::Ptr > bunny_clouds_;
-    std::vector<Eigen::Matrix4d> pairwise_transformations_;
+    std::vector<Eigen::Matrix4f> pairwise_transformations_;
     transducer dq_transducer;
 
     void loadPCDFiles();
     void saveCloudKeyPoints(std::string, int);
-    void getTransformOfPair(pcl::PointCloud<Point>::Ptr& cloud_1, pcl::PointCloud<Point>::Ptr& cloud_2, Eigen::Matrix4d& tranform);
+    void getTransformOfPair(pcl::PointCloud<Point>::Ptr& cloud_1, pcl::PointCloud<Point>::Ptr& cloud_2, Eigen::Matrix4f& tranform);
 
 };
